Move ReviewPolicyClass to Review table
