+++ app/components/task_list_items/assessment/policy_class_component.rb
@@ -25,10 +25,22 @@ module TaskListItems
       end
 
       def status_tag_component
-        StatusTags::PolicyClassComponent.new(
-          policy_class:,
-          planning_application:
-        )
+        StatusTags::BaseComponent.new(status:)
+      end
+
+      def status
+        if to_be_reviewed?
+          :to_be_reviewed
+        elsif policy_class.in_assessment?
+          :in_progress
+        elsif policy_class.complete?
+          :complete
+        end
+      end
+
+      def to_be_reviewed?
+        planning_application.recommendation&.rejected? &&
+          policy_class.update_required?
       end
     end
   end
