+++ app/views/planning_applications/assessment/policy_classes/_summary.html.erb
@@ -23,13 +23,13 @@
       <%= policy_class.name.upcase_first %>
     </h2>
 
-    <% if policy_class&.review_policy_class&.comment %>
+    <% if policy_class&.review&.comment %>
       <div class="govuk-inset-text" id="reviewer_comment">
         <p class="govuk-body govuk-!-font-weight-bold govuk-!-margin-bottom-0">
           Reviewer comment:
         </p>
-        <p class="govuk-body govuk-!-margin-top-1"><%= policy_class.review_policy_class.created_at.to_fs %></p>
-        <p class="govuk-body italic"><%= policy_class.review_policy_class.comment %></p>
+        <p class="govuk-body govuk-!-margin-top-1"><%= policy_class.review.created_at.to_fs %></p>
+        <p class="govuk-body italic"><%= policy_class.review.comment %></p>
       </div>
     <% end %>
     <%= button_to(
